update need to fix for resign date

+++ src/main/java/dev/freaks/BADProject02/service/EmployeeService.java
@@ -10,8 +10,8 @@ import java.util.List;
 public interface EmployeeService {
     List<EmployeeResponseDto> getAllEmployees();
     EmployeeResponseDto createEmployee(EmployeeCreateDto employeeCreateDto);
-    EmployeeResponseDto getEmployeeById(Long id);
-    EmployeeResponseDto updateEmployee(Long id, EmployeeUpdateDto employeeUpdateDto);
-    void deleteEmployee(Long id);
+    EmployeeResponseDto getEmployeeById(Integer id);
+    EmployeeResponseDto updateEmployee(Integer id, EmployeeUpdateDto employeeUpdateDto);
+    void deleteEmployee(Integer id);
     List<TopPaidEmployeeDto> getTop10HighestPaidEmployees();
 }
